Adding .pre-commit-config.yaml & minor improvments

+++ jsonp_db-tests.sh
@@ -4,7 +4,7 @@ DB_FILE=jsonp_db.db
 set -o pipefail -o errexit -o nounset -o xtrace
 
 make kill-local-server
-mv $DB_FILE $DB_FILE.bak
+test -e $DB_FILE && mv $DB_FILE $DB_FILE.bak
 sqlite3 $DB_FILE 'CREATE TABLE KVStore(Key TEXT PRIMARY KEY, Value TEXT);'
 make start-local-server
 
@@ -13,7 +13,7 @@ curl -s http://localhost:8080/jsonp_db/the_answer | grep '^42$'
 curl -sX POST -d '{name:"John Doe"}' http://localhost:8080/jsonp_db/json_doe | grep '{name:"John Doe"}'
 curl -s http://localhost:8080/jsonp_db/json_doe | grep '{name:"John Doe"}'
 curl -s http://localhost:8080/jsonp_db/urlencoded_dict?%7Bname%3A%22John%20Doe%22%7D | grep '{name:"John Doe"}'
-curl -s http://localhost:8080/jsonp_db/urlencoded_dict?callback=foo | grep 'foo({name:"John Doe"}, '
+curl -s http://localhost:8080/jsonp_db/urlencoded_dict?callback=foo | grep 'foo({name:"John Doe"}, null)'
 curl -s http://localhost:8080/jsonp_db/nested?%7Ba%3A%7Bb%3Atrue%7D%7D | grep '{a:{b:true}}'
 echo '@<>#%"{}|\^[]`' > tmp.json
 curl -sX POST --data-urlencode @tmp.json http://localhost:8080/jsonp_db/urlencoded_str | grep '@<>#%"{}|\\^\[\]`'
@@ -29,8 +29,12 @@ curl -sX PUT -d 0=1 http://localhost:8080/jsonp_db/$RANDOM_101CHAR_KEY | grep 'K
 # TODO FIXME
 
 # Modification-key
-# TODO FIXME
+modifkey=$(curl -s 'http://localhost:8080/jsonp_db/K?callback=_&V1' | sed 's/_(V1, "\(.*\)")/\1/')
+curl -s "http://localhost:8080/jsonp_db/K?V2&modification-key=$modifkey"
+curl -s http://localhost:8080/jsonp_db/K | grep V2
+curl -s 'http://localhost:8080/jsonp_db/K?V2&modification-key=DUMMY' | grep 'Error400: Invalid modification-key'
+curl -s http://localhost:8080/jsonp_db/K?V2 | grep 'Error400: No modification-key provided'
 
 make kill-local-server
-mv $DB_FILE.bak $DB_FILE
+test -e $DB_FILE.bak && mv $DB_FILE.bak $DB_FILE
